Update minimum significance value
After speaking with others, 10e-100 seems like a better minimum significance value.
In order to use such a precise value significance has to be a float64.

+++ src/merge.rs
@@ -176,14 +176,14 @@ pub fn merge_filtered_data(
     let numeric_intervals = result_data.iter().map(|d| d.numeric_intervals).fold(
         FilterIntervals {
             effect: (f32::MAX, f32::MIN),
-            sig: (f32::MAX, f32::MIN),
+            sig: (f64::MAX, f64::MIN),
         },
         |acc, d| FilterIntervals {
             effect: (
                 f32::min(acc.effect.0, d.effect.0),
                 f32::max(acc.effect.1, d.effect.1),
             ),
-            sig: (f32::min(acc.sig.0, d.sig.0), f32::max(acc.sig.1, d.sig.1)),
+            sig: (f64::min(acc.sig.0, d.sig.0), f64::max(acc.sig.1, d.sig.1)),
         },
     );
